refactor(events): unify commit and file change types in FileChangeKind enum
Removed obsolete CommitEventType enum and merged its values into FileChangeKind
to consolidate all commit-related changes under a single type, simplifying event
handling and type safety across TraceEvent and TraceEventFactory.
