--- src/Core/Enums/CommitEventType.cs
@@ -1,13 +1,0 @@
-namespace ChangeTrace.Core.Enums;
-
-/// <summary>
-/// Represents type of commit related event in repository.
-/// </summary>
-internal enum CommitEventType
-{
-    Commit,
-    FileAdded,
-    FileModified,
-    FileDeleted,
-    FileRenamed
-}
